Split up config.ini further
