+++ DTXMania/Core/CConfigIni.cs
@@ -10,466 +10,6 @@ internal partial class CConfigIni
 {
 	// Class
 
-	public class CKeyAssign
-	{
-		public class CKeyAssignPad
-		{
-			public STKEYASSIGN[] HH
-			{
-				get => padHH_R;
-				set => padHH_R = value;
-			}
-
-			public STKEYASSIGN[] R
-			{
-				get => padHH_R;
-				set => padHH_R = value;
-			}
-
-			public STKEYASSIGN[] SD
-			{
-				get => padSD_G;
-				set => padSD_G = value;
-			}
-
-			public STKEYASSIGN[] G
-			{
-				get => padSD_G;
-				set => padSD_G = value;
-			}
-
-			public STKEYASSIGN[] BD
-			{
-				get => padBD_B;
-				set => padBD_B = value;
-			}
-
-			public STKEYASSIGN[] B
-			{
-				get => padBD_B;
-				set => padBD_B = value;
-			}
-
-			public STKEYASSIGN[] HT
-			{
-				get => padHT_Pick;
-				set => padHT_Pick = value;
-			}
-
-			public STKEYASSIGN[] Pick
-			{
-				get => padHT_Pick;
-				set => padHT_Pick = value;
-			}
-
-			public STKEYASSIGN[] LT
-			{
-				get => padLT_Wail;
-				set => padLT_Wail = value;
-			}
-
-			public STKEYASSIGN[] Wail
-			{
-				get => padLT_Wail;
-				set => padLT_Wail = value;
-			}
-
-			public STKEYASSIGN[] FT
-			{
-				get => padFT_Help;
-				set => padFT_Help = value;
-			}
-
-			public STKEYASSIGN[] Help
-			{
-				get => padFT_Help;
-				set => padFT_Help = value;
-			}
-
-			public STKEYASSIGN[] CY
-			{
-				get => padCY_Decide;
-				set => padCY_Decide = value;
-			}
-
-			public STKEYASSIGN[] Decide
-			{
-				get => padCY_Decide;
-				set => padCY_Decide = value;
-			}
-
-			public STKEYASSIGN[] HHO
-			{
-				get => padHHO_Y;
-				set => padHHO_Y = value;
-			}
-
-			public STKEYASSIGN[] Y
-			{
-				get => padHHO_Y;
-				set => padHHO_Y = value;
-			}
-
-			public STKEYASSIGN[] RD
-			{
-				get => padRD;
-				set => padRD = value;
-			}
-
-			public STKEYASSIGN[] P
-			{
-				get => padLC_P;
-				set => padLC_P = value;
-			}
-
-			public STKEYASSIGN[] LC
-			{
-				get => padLC_P;
-				set => padLC_P = value;
-			}
-
-			public STKEYASSIGN[] LP
-			{
-				get => padLP;
-				set => padLP = value;
-			}
-
-			public STKEYASSIGN[] LBD
-			{
-				get => padLBD;
-				set => padLBD = value;
-			}
-
-			public STKEYASSIGN[] Cancel
-			{
-				get => padCancel;
-				set => padCancel = value;
-			}
-
-			public STKEYASSIGN[] Capture
-			{
-				get => padCapture;
-				set => padCapture = value;
-			}
-
-			public STKEYASSIGN[] Search
-			{
-				get => padSearch;
-				set => padSearch = value;
-			}
-
-			public STKEYASSIGN[] LoopCreate
-			{
-				get => padLoopCreate;
-				set => padLoopCreate = value;
-			}
-
-			public STKEYASSIGN[] LoopDelete
-			{
-				get => padLoopDelete;
-				set => padLoopDelete = value;
-			}
-
-			public STKEYASSIGN[] SkipForward
-			{
-				get => padSkipForward;
-				set => padSkipForward = value;
-			}
-
-			public STKEYASSIGN[] SkipBackward
-			{
-				get => padSkipBackward;
-				set => padSkipBackward = value;
-			}
-
-			public STKEYASSIGN[] IncreasePlaySpeed
-			{
-				get => padIncreasePlaySpeed;
-				set => padIncreasePlaySpeed = value;
-			}
-
-			public STKEYASSIGN[] DecreasePlaySpeed
-			{
-				get => padDecreasePlaySpeed;
-				set => padDecreasePlaySpeed = value;
-			}
-
-			public STKEYASSIGN[] Restart
-			{
-				get => padRestart;
-				set => padRestart = value;
-			}
-
-			public STKEYASSIGN[] this[int index]
-			{
-				get
-				{
-					switch (index)
-					{
-						case (int)EKeyConfigPad.HH:
-							return padHH_R;
-
-						case (int)EKeyConfigPad.SD:
-							return padSD_G;
-
-						case (int)EKeyConfigPad.BD:
-							return padBD_B;
-
-						case (int)EKeyConfigPad.HT:
-							return padHT_Pick;
-
-						case (int)EKeyConfigPad.LT:
-							return padLT_Wail;
-
-						case (int)EKeyConfigPad.FT:
-							return padFT_Help;
-
-						case (int)EKeyConfigPad.CY:
-							return padCY_Decide;
-
-						case (int)EKeyConfigPad.HHO:
-							return padHHO_Y;
-
-						case (int)EKeyConfigPad.RD:
-							return padRD;
-
-						case (int)EKeyConfigPad.LC:
-							return padLC_P;
-
-						case (int)EKeyConfigPad.LP: // #27029 2012.1.4 from
-							return padLP; //(HPからLPに。)
-
-						case (int)EKeyConfigPad.LBD:
-							return padLBD;
-
-						case (int)EKeyConfigPad.Cancel:
-							return padCancel;
-
-						case (int)EKeyConfigPad.Capture:
-							return padCapture;
-
-						case (int)EKeyConfigPad.Search:
-							return padSearch;
-
-						case (int)EKeyConfigPad.LoopCreate:
-							return padLoopCreate;
-
-						case (int)EKeyConfigPad.LoopDelete:
-							return padLoopDelete;
-
-						case (int)EKeyConfigPad.SkipForward:
-							return padSkipForward;
-
-						case (int)EKeyConfigPad.SkipBackward:
-							return padSkipBackward;
-
-						case (int)EKeyConfigPad.IncreasePlaySpeed:
-							return padIncreasePlaySpeed;
-
-						case (int)EKeyConfigPad.DecreasePlaySpeed:
-							return padDecreasePlaySpeed;
-
-						case (int)EKeyConfigPad.Restart:
-							return padRestart;
-					}
-
-					throw new IndexOutOfRangeException();
-				}
-				set
-				{
-					switch (index)
-					{
-						case (int)EKeyConfigPad.HH:
-							padHH_R = value;
-							return;
-
-						case (int)EKeyConfigPad.SD:
-							padSD_G = value;
-							return;
-
-						case (int)EKeyConfigPad.BD:
-							padBD_B = value;
-							return;
-
-						case (int)EKeyConfigPad.Pick:
-							padHT_Pick = value;
-							return;
-
-						case (int)EKeyConfigPad.LT:
-							padLT_Wail = value;
-							return;
-
-						case (int)EKeyConfigPad.FT:
-							padFT_Help = value;
-							return;
-
-						case (int)EKeyConfigPad.CY:
-							padCY_Decide = value;
-							return;
-
-						case (int)EKeyConfigPad.HHO:
-							padHHO_Y = value;
-							return;
-
-						case (int)EKeyConfigPad.RD:
-							padRD = value;
-							return;
-
-						case (int)EKeyConfigPad.LC:
-							padLC_P = value;
-							return;
-
-						case (int)EKeyConfigPad.LP:
-							padLP = value;
-							return;
-
-						case (int)EKeyConfigPad.LBD:
-							padLBD = value;
-							return;
-
-						case (int)EKeyConfigPad.Cancel:
-							padCancel = value;
-							return;
-
-						case (int)EKeyConfigPad.Capture:
-							padCapture = value;
-							return;
-
-						case (int)EKeyConfigPad.Search:
-							padSearch = value;
-							return;
-
-						case (int)EKeyConfigPad.LoopCreate:
-							padLoopCreate = value;
-							return;
-
-						case (int)EKeyConfigPad.LoopDelete:
-							padLoopDelete = value;
-							return;
-
-						case (int)EKeyConfigPad.SkipForward:
-							padSkipForward = value;
-							return;
-
-						case (int)EKeyConfigPad.SkipBackward:
-							padSkipBackward = value;
-							return;
-
-						case (int)EKeyConfigPad.IncreasePlaySpeed:
-							padIncreasePlaySpeed = value;
-							return;
-
-						case (int)EKeyConfigPad.DecreasePlaySpeed:
-							padDecreasePlaySpeed = value;
-							return;
-
-						case (int)EKeyConfigPad.Restart:
-							padRestart = value;
-							return;
-					}
-
-					throw new IndexOutOfRangeException();
-				}
-			}
-
-			#region [ private ]
-
-			//-----------------
-			private STKEYASSIGN[] padBD_B;
-			private STKEYASSIGN[] padCY_Decide;
-			private STKEYASSIGN[] padFT_Help;
-			private STKEYASSIGN[] padHH_R;
-			private STKEYASSIGN[] padHHO_Y;
-			private STKEYASSIGN[] padHT_Pick;
-			private STKEYASSIGN[] padLC_P;
-			private STKEYASSIGN[] padLT_Wail;
-			private STKEYASSIGN[] padRD;
-			private STKEYASSIGN[] padSD_G;
-			private STKEYASSIGN[] padLP;
-			private STKEYASSIGN[] padLBD;
-			private STKEYASSIGN[] padCancel;
-			private STKEYASSIGN[] padCapture;
-			private STKEYASSIGN[] padSearch;
-			private STKEYASSIGN[] padLoopCreate;
-			private STKEYASSIGN[] padLoopDelete;
-			private STKEYASSIGN[] padSkipForward;
-			private STKEYASSIGN[] padSkipBackward;
-			private STKEYASSIGN[] padIncreasePlaySpeed;
-			private STKEYASSIGN[] padDecreasePlaySpeed;
-
-			private STKEYASSIGN[] padRestart;
-			//-----------------
-
-			#endregion
-		}
-
-		[StructLayout(LayoutKind.Sequential)]
-		public struct STKEYASSIGN
-		{
-			public EInputDevice InputDevice;
-			public int ID;
-			public int Code;
-
-			public STKEYASSIGN(EInputDevice DeviceType, int nID, int nCode)
-			{
-				InputDevice = DeviceType;
-				ID = nID;
-				Code = nCode;
-			}
-		}
-
-		public CKeyAssignPad Bass = new CKeyAssignPad();
-		public CKeyAssignPad Drums = new CKeyAssignPad();
-		public CKeyAssignPad Guitar = new CKeyAssignPad();
-		public CKeyAssignPad System = new CKeyAssignPad();
-
-		public CKeyAssignPad this[int index]
-		{
-			get
-			{
-				switch (index)
-				{
-					case (int)EKeyConfigPart.DRUMS:
-						return Drums;
-
-					case (int)EKeyConfigPart.GUITAR:
-						return Guitar;
-
-					case (int)EKeyConfigPart.BASS:
-						return Bass;
-
-					case (int)EKeyConfigPart.SYSTEM:
-						return System;
-				}
-
-				throw new IndexOutOfRangeException();
-			}
-			set
-			{
-				switch (index)
-				{
-					case (int)EKeyConfigPart.DRUMS:
-						Drums = value;
-						return;
-
-					case (int)EKeyConfigPart.GUITAR:
-						Guitar = value;
-						return;
-
-					case (int)EKeyConfigPart.BASS:
-						Bass = value;
-						return;
-
-					case (int)EKeyConfigPart.SYSTEM:
-						System = value;
-						return;
-				}
-
-				throw new IndexOutOfRangeException();
-			}
-		}
-	}
-
 	public enum ESoundDeviceTypeForConfig
 	{
 		ACM = 0,
@@ -1295,14 +835,7 @@ internal partial class CConfigIni
 		tWriteSchemaSection( sw, "DiscordRichPresence" );
 		sw.WriteLine( ";-------------------" );
 
-		// [GUID] uses repeated JoystickID keys, so it is not a simple key=value section.
-		sw.WriteLine( "[GUID]" );
-		sw.WriteLine();
-		foreach( KeyValuePair<int, string> pair in joystickDict )
-		{
-			sw.WriteLine( "JoystickID={0},{1}", pair.Key, pair.Value );
-		}
-		sw.WriteLine();
+		tWriteSchemaSection( sw, "GUID" );
 
 		// The key-assign legend is emitted as [DrumsKeyAssign]'s header lines (see KeyAssignLegend).
 		tWriteSchemaSection( sw, "DrumsKeyAssign" );
@@ -1362,17 +895,9 @@ internal partial class CConfigIni
 				string str3 = strArray[ 0 ].Trim();
 				string str4 = strArray[ 1 ].Trim();
 
-				// Almost every section is described declaratively by the schema (CConfigIni.Schema.cs)
-				// and dispatched through a single lookup. [GUID] is the only exception: it uses repeated
-				// JoystickID keys, so it cannot be modelled as one key -> one value.
-				if ( bIsSchemaSection( currentSectionName ) )
-				{
-					tReadSchemaField( currentSectionName, str3, str4 );
-				}
-				else if ( currentSectionName.Equals( "GUID" ) && str3.Equals( "JoystickID" ) )
-				{
-					tAcquireJoystickID( str4 );
-				}
+				// Every section is described declaratively by the schema (CConfigIni.Schema.cs) and
+				// dispatched through a single lookup — no per-section special cases.
+				tReadSchemaField( currentSectionName, str3, str4 );
 			}
 			catch ( Exception exception )
 			{
@@ -1430,18 +955,8 @@ internal partial class CConfigIni
 	}
 	private void tClearAllKeyAssignments()
 	{
+		// A fresh CKeyAssign already allocates every pad and fills it with unassigned (Unknown) slots.
 		KeyAssign = new CKeyAssign();
-		for( int i = 0; i <= (int)EKeyConfigPart.SYSTEM; i++ )
-		{
-			for( int j = 0; j < (int)EKeyConfigPad.MAX; j++ )
-			{
-				KeyAssign[ i ][ j ] = new CKeyAssign.STKEYASSIGN[ 16 ];
-				for( int k = 0; k < 16; k++ )
-				{
-					KeyAssign[ i ][ j ][ k ] = new CKeyAssign.STKEYASSIGN( EInputDevice.Unknown, 0, 0 );
-				}
-			}
-		}
 	}
 	private void tReadAndSetSkey(string strキー記述, CKeyAssign.STKEYASSIGN[] assign)
 	{
